Improved the error handling for duplicate ids.
The issue was that undeclared identifiers were being reported several times,
for the same identifier. The present system on errors was to keep a set of
"unknown matching entries", which are identifer structures filled out with
general fields, a type link that is nil, meaning don't produce matching errors
on this entry, and is treated as a wildcard in type comparisions.

But every reference to a undefined identifier gives the same entry, and it
is not entered into the identifiers for the current block. So errors
are redundant. In a simple fix, we move those "undefineds" into the searchid()
routine, and give them a name for each undefined. In other words, previously
all of the error references give the same entry, which is anonymous. Now
each error reference gets a different entry, that bears the undefined name.

The net result of this is that only the first error, "identifier not declared",
is output and all of the rest of the errors on that same entry are supressed.
It even still tracks thing like "never referenced" tracking, even though
the identifier was never declared.

The drawback of the new system is that it creates more ids. However, note that
this only happens in the presence of errors, and only creates one entry, and
that entry would have been created in any case if the identifier had be
declared. Its more like saying that the errors case saved on memory, vs. saying
we added new memory loading.

 Changes to be committed:
       modified:   regress_report.txt
       modified:   source/pcom.pas

1. Removed undeclared id entrys that are no longer used. Vars, procedures and
functions still have an undeclared id set up for them, because at various
points in the code they are used to solve various error recovery cases.

2. In searchid(), instead of picking up a global undeclared id entry,
we make one up and enter it with enterid(). This is the same code used in
enterundecl() moved up.

3. In enterundecl() and exitundecl() we remove the above entries.