Normalized on Unix/linux line endings.
=== Files changed ===

	modified:   p2/roman.cmp
	modified:   p4/standardp.cmp
	modified:   sample_programs/basics.cmp
	modified:   sample_programs/drystone.cmp
	modified:   sample_programs/fbench.cmp
	modified:   sample_programs/hello.cmp
	modified:   sample_programs/match.cmp
	modified:   sample_programs/pascals.cmp
	modified:   sample_programs/prime.cmp
	modified:   sample_programs/qsort.cmp
	modified:   sample_programs/roman.cmp
	modified:   sample_programs/startrek.cmp
	modified:   source/pcom.pas
	modified:   source/pint.pas
	modified:   standard_tests/iso7185pat.cmp
	modified:   standard_tests/iso7185prt1707a.cmp
	modified:   standard_tests/iso7185prt1707a.pas
	modified:   standard_tests/iso7185prt1707b.cmp
	modified:   standard_tests/iso7185prt1707b.pas
	modified:   standard_tests/iso7185prt1728.cmp
	modified:   standard_tests/iso7185prt1731.cmp
	modified:   standard_tests/iso7185prt1732.cmp
	modified:   standard_tests/iso7185prt1737.cmp
	modified:   standard_tests/iso7185prt1751.cmp
	modified:   standard_tests/iso7185prt1755.cmp
	modified:   standard_tests/iso7185prt1828.cmp
	modified:   standard_tests/iso7185prt1901.pas

All these were normalized to Unix/linux line endings. This should
tamp down the git warnings.

=== Bug numbers affected ===

=== Notes ===

In case it is not evident, I really hate the line ending wars between
Unix and the rest of the world. The answer is don't build applications
that care! It is SOOO simple to create programs that are line ending
agnostic, and even tolerate mixed line endings. All the programs I
have control over do this (see flip for a good example). The issue
is that a lot of older, crufty Unix/GNU/Linux programs have not been
updated, like for example bash, and this is always a fight.